ai_project/requirements.txt
streamlit
faiss-cpu
sentence-transformers
torch
numpy
PyPDF2
